Keep CSP nonce helper on the binary-safe framework path
