+++ src/react/components/ai/csp-nonce.ts
@@ -1,1 +1,13 @@
-export { getDocumentNonce } from "../csp-nonce.ts";
+/**
+ * Reuse the server-issued CSP nonce for client-created style/script elements
+ * during hydration and SPA updates.
+ */
+export function getDocumentNonce(): string | undefined {
+  if (typeof document === "undefined") return undefined;
+
+  const element = document.querySelector<HTMLElement>("script[nonce], style[nonce], link[nonce]");
+  if (!element) return undefined;
+
+  const nonce = element.nonce || element.getAttribute("nonce") || "";
+  return nonce || undefined;
+}
