--- src/react/components/csp-nonce.ts
@@ -1,13 +1,0 @@
-/**
- * Reuse the server-issued CSP nonce for client-created style/script elements
- * during hydration and SPA updates.
- */
-export function getDocumentNonce(): string | undefined {
-  if (typeof document === "undefined") return undefined;
-
-  const element = document.querySelector<HTMLElement>("script[nonce], style[nonce], link[nonce]");
-  if (!element) return undefined;
-
-  const nonce = element.nonce || element.getAttribute("nonce") || "";
-  return nonce || undefined;
-}
